ekleme örneği ve post filtreleri
