+++ kodilan/mail.py
@@ -8,7 +8,7 @@ def do_mail(email, title, theme, variables):
     msg_html = render_to_string('mail/{}.html'.format(theme), variables)
     try:
         send_mail(
-            'Oleybee - ' + title,
+            'Kodilan - ' + title,
             msg_plain,
             getattr(settings, "SENDER_MAIL", ""),
             [email],
@@ -20,6 +20,7 @@ def do_mail(email, title, theme, variables):
     pass
 
 
+# todo: create post will update
 def send_activation(email, first_name, last_name, token):
     do_mail(email, "İlan Onay", "activation",
             {'first_name': first_name, 'last_name': last_name, 'code': token})
