+++ kodilan/templates/mail/activation.txt
