+++ posts/views.py
@@ -1,13 +1,11 @@
 from .models import Post, Company, Tag
-from .serializers import PostSerializer, CompanySerializer, TagSerializer
+from .serializers import PostSerializer, CompanySerializer, TagSerializer, CreatePostSerializer
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from rest_framework.generics import ListAPIView
+from rest_framework.generics import ListAPIView, CreateAPIView
 from rest_framework import filters
 from django_filters.rest_framework import DjangoFilterBackend
 from .filters import PeriodFilterBackend, StatusFilterBackend
-from rest_framework import pagination
-from core.customs import StandardResultsSetPagination
 
 
 class CompaniesView(ListAPIView):
@@ -16,7 +14,6 @@ class CompaniesView(ListAPIView):
     filter_backends = [filters.SearchFilter, filters.OrderingFilter]
     search_fields = ['name']
     ordering_fields = ['id', 'name', 'created_at']
-    pagination_class = StandardResultsSetPagination
 
 
 class PostsView(ListAPIView):
@@ -27,8 +24,12 @@ class PostsView(ListAPIView):
 
     search_fields = ['is_featured', 'location', 'type', 'position', 'title', 'tags']
     ordering_fields = ['pub_date', 'created_at']
-    pagination_class = StandardResultsSetPagination
     filterset_fields = ['category', 'is_finish']
+
+
+class CreatePostsView(CreateAPIView):
+    queryset = Post.objects.all()
+    serializer_class = CreatePostSerializer
 
 
 class TagsView(ListAPIView):
@@ -36,18 +37,12 @@ class TagsView(ListAPIView):
     serializer_class = TagSerializer
     filter_backends = [filters.SearchFilter, filters.OrderingFilter]
     search_fields = ['name']
-    pagination_class = StandardResultsSetPagination
 
 
 class FindLocationAction(APIView):
     def get(self, request):
         locations = Post.objects.all().distinct('location')
-
         data = {}
-
         for location in locations:
             data[location.id] = location.location
-
         return Response(data)
-
-        return queryset
